+++ models/transformations/base/t_fibre.sql
@@ -10,7 +10,7 @@
 }}
 
 SELECT
-    row_number() OVER (ORDER BY fo_code) AS id,
+    row_number() OVER (ORDER BY fo_code)::int4 AS id,
     CASE WHEN pg_input_is_valid(NULLIF(fo_cb_code::text, ''), 'varchar(254)') THEN fo_cb_code::VARCHAR(254) ELSE NULL END AS fo_cb_code,
     CASE WHEN pg_input_is_valid(NULLIF(fo_code::text, ''), 'varchar(254)') THEN fo_code::VARCHAR(254) ELSE NULL END AS fo_code,
     CASE WHEN pg_input_is_valid(NULLIF(fo_etat::text, ''), 'varchar(3)') THEN fo_etat::VARCHAR(3) ELSE NULL END AS fo_etat,
